Adding strtok + strspn + strcspn + strpbrk and updating previous implementations

+++ string_length/string_length.h
@@ -3,6 +3,8 @@
 #ifndef STRING_LENGTH
 #define STRING_LENGTH
 
-unsigned int string_length(const char* p_string);
+#include <stddef.h>
+
+size_t string_length(const char* p_string);
 
 #endif
